requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bleach==6.1.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
webencodings==0.5.1
drf_yasg